+++ docs/html/annotated_dup.js
@@ -4,7 +4,6 @@ var annotated_dup =
     [ "BmsDeviceReader", "class_bms_device_reader.html", null ],
     [ "BmsDeviceUpdater", "class_bms_device_updater.html", null ],
     [ "BmsDeviceWriter", "class_bms_device_writer.html", null ],
-    [ "RaceUp_ino_core", "class_race_up__ino__core.html", null ],
     [ "SdCardStream", "class_sd_card_stream.html", "class_sd_card_stream" ],
     [ "Utils", "class_utils.html", null ]
 ];