+++ docs/html/search/all_1.js
@@ -1,5 +1,6 @@
 var searchData=
 [
+  ['celsius_5fto_5fkelvin_5ffactor',['CELSIUS_TO_KELVIN_FACTOR',['../class_utils.html#af916f41aa01c118944d9dd91fea3584e',1,'Utils']]],
   ['clearalertsonbmsdevice',['clearAlertsOnBmsDevice',['../class_bms_device.html#a6f8832d9d4211aca63fae77444262a16',1,'BmsDevice']]],
   ['clearfaultsonbmsdevice',['clearFaultsOnBmsDevice',['../class_bms_device.html#a26155b471a97e1bd96485f2e95934a11',1,'BmsDevice']]],
   ['commandtostring',['commandToString',['../class_bms_device_updater.html#aff1720a11582bd1353c50ced9352eda9',1,'BmsDeviceUpdater']]],