+++ docs/html/search/functions_3.js
@@ -1,4 +1,5 @@
 var searchData=
 [
+  ['frametofloat',['frameToFloat',['../class_utils.html#af819499b6a69f2aed478a2945bedc56b',1,'Utils::frameToFloat(CAN_FRAME &amp;frame, float *x, float *x1)'],['../class_utils.html#aa3f908c62401cc150a33ac791e0fa1c1',1,'Utils::frameToFloat(CAN_FRAME &amp;frame, float *x)']]],
   ['fromkelvintocelsius',['fromKelvinToCelsius',['../class_utils.html#a393f44e0230724e5cd3b7c125b16b5e1',1,'Utils']]]
 ];